Fix FaceEnrollPredictResult ambigious structure and add the field code
The field 'code'   is a code or error value returned of the operation, this is needed t distinguish between what is returned by a successful enroll or predict calls from the success of the call itself. A call is successful if it returned an intelligible JSON.
TODO: 1) the field 'code '  to be split in 2 fields after the fix the spec in c/c++ code
TODO: 2) pickup a proper name like to express the nature of these codes like error_code, validation_code etc.